+++ app/Http/Controllers/EventController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Classroom;
 use App\Event;
+use Arcanedev\LaravelSettings\Utilities\Arr;
 use Illuminate\Http\Request;
 
 class EventController extends Controller
@@ -13,22 +14,105 @@ class EventController extends Controller
         $this->middleware('verified');
     }
 
+    public function index($code)
+    {
+        $class = Classroom::where('code', '=', $code)->firstOrFail();
+        settings()->setExtraColumns(['classroom_id' => $class->id]);
+        $this->authorize('view', $class);
+
+        $events = Event::where('type', "!=", 13)->whereNull('classroom_id')->orWhere('classroom_id', '=', $class->id)->get();
+        $disabled = settings()->get('disabled_events', json_encode([]));
+        
+        return view('events.index',  compact('class', 'events', 'disabled'));
+    }
+    
+    public function update()
+    {
+        $data = request()->validate([
+            'event.id' => ['numeric', 'required'],
+            'event.title' => ['string', 'required'],
+            'event.content' => ['string', 'required'],
+            'event.type' => ['numeric', 'required'],
+            'event.classroom_id' => ['numeric', 'required'],
+        ]);
+    
+        $event = Event::findOrFail($data['event']['id']);
+        $class = Classroom::findOrFail($event->classroom_id);
+        $this->authorize('update', $class);
+    
+        return $event->update($data['event']);
+    }
+
+    public function add()
+    {
+        $data = request()->validate([
+            'event.title' => ['string', 'required'],
+            'event.content' => ['string', 'required'],
+            'event.type' => ['numeric', 'required'],
+            'event.classroom_id' => ['numeric', 'required'],
+        ]);
+
+        $class = Classroom::findOrFail($data['event']['classroom_id']);
+        $this->authorize('update', $class);
+
+        return Event::create($data['event']);
+    }
+
+    public function disable($code)
+    {
+        $class = Classroom::where('code', '=', $code)->firstOrFail();
+        $this->authorize('update', $class);
+        settings()->setExtraColumns(['classroom_id' => $class->id]);
+
+        $data = request()->validate([
+            'id' => ['numeric', 'required']
+        ]);
+
+        $disabled = json_decode(settings()->get('disabled_events', json_encode([])));
+
+        $exists = array_search($data['id'], $disabled); 
+
+        if($exists !== false) {
+            array_splice($disabled, $exists, 1);
+        } else {
+            array_push($disabled, $data['id']);
+        }
+
+        settings()->set('disabled_events', json_encode($disabled));
+
+    }
+
+    public function destroy($id)
+    {
+        $event = Event::where('id', '=', $id)->first();
+        $class = Classroom::where('id', $event->classroom_id)->firstOrFail();
+        $this->authorize('update', $class);
+        try {
+            $event->delete();
+        } catch (\Throwable $th) {
+            return ['error' => $th];
+        }
+        return 1;
+    }
+    
     public function show($code)
     {
         $class = Classroom::where('code', '=', $code)->firstOrFail();
         $this->authorize('view', $class);
         $max = 0;
         settings()->setExtraColumns(['classroom_id' => $class->id]);
-        $disabled = collect(settings()->get('disabled_events'));
+        $disabled = json_decode(settings()->get('disabled_events', "[]"));
 
         do {
             $again = false;
-            $event = Event::whereNull('classroom_id')->orWhere('classroom_id', '=', $class->id)->inRandomOrder()->first();
-            if($disabled->contains($event->id)) {
-                $again = true;
-            } else if (($event->type == 1 || $event->type == 3 || $event->type == 4 || $event->type == 8 || $event->type == 9 || $event->type == 11 || $event->type == 12 || $event->type == 13) && !count($class->students)) {
-                $again = true;
-            } else if (($event->type == 2 || $event->type == 10) && !count($class->grouping->first()->groups)) {
+            $event = Event::whereNotIn('id', $disabled)
+                            ->where(function($query) use ($class){
+                                $query->whereNull('classroom_id');
+                                $query->orWhere('classroom_id', '=', $class->id);
+                            })
+                            ->inRandomOrder()
+                            ->first();
+            if (($event->type == 1 || $event->type == 3 || $event->type == 2 || $event->type == 4 || $event->type == 8 || $event->type == 9 || $event->type == 10 || $event->type == 11 || $event->type == 12 || $event->type == 13) && !count($class->students)) {
                 $again = true;
             } else if ($event->type == 5 && (!$class->students || !count($class->grouping->first()->groups))) {
                 $again = true;
@@ -41,9 +125,11 @@ class EventController extends Controller
                 $again = true;
             }
             $max++;
-        } while ($again && $max < 100);
+        } while ($again && $max < 1000);
 
         $return = null;
+        if($max == 1000)
+            return redirect('/classroom/'.$class->code.'/');
         switch ($event->type) {
             case 0:
                 $return = [
@@ -62,7 +148,6 @@ class EventController extends Controller
                 $options = json_decode($event->options);
                 $return = [
                     'hp' => rand($options->hpMin, $options->hpMax) * -1,
-                    'group' => GroupsController::getRandomGroup($class),
                     'event' => $event,
                 ];
                 break;