+++ resources/js/app.js
@@ -97,6 +97,7 @@ Vue.component('create-map', require('./components/map/CreateMap.vue').default);
 Vue.component('settings', require('./components/settings/Settings.vue').default);
 
 Vue.component('show-event', require('./components/event/ShowEvent.vue').default);
+Vue.component('show-events', require('./components/event/ShowEvents.vue').default);
 
 Vue.component('edit-rules', require('./components/rules/EditRules.vue').default);