+++ resources/lang/en/general.php
@@ -27,6 +27,9 @@ return [
     
     'no' => "No",
 
+    'enable' => "Enable",
+    'disable' => "Disable",
+
     'confirm_delete' => "Are you sure you want to delete this element? This action can't be undone.",
     'confirm_delete_class' => "Are you sure you want to delete this element? This action can't be undone. THIS WILL DESTROY ALL THE STUDENT INFORMATION.",
     'delete' => "Delete",