+++ resources/views/events/index.blade.php
@@ -1,0 +1,5 @@
+@extends('layouts.teacher')
+
+@section('content')
+    <show-events :events="{{ $events }}" :classroom="{{ $class }}" :disabled="{{ $disabled }}"></show-events>
+@endsection